Version 1.0.2
Uses MessageContent class to handle text colouring server-side rather than attempting to determine the colour of text client-side.
The server creates a MessageContent object with variables "text" and "colour". This object is then put in a list where text used to be. This list is sent to the each client - each client can then write the message to each line with the colour specified.

This optimises code as before, a possible 36 if statements (3 if statements for each of 12 message boxes) were called per message sent, per client - this could have gotten slow on older hardware. These if statements tested whether a colon was present in the text, and if so, the message would be black, as it would be a standard message. If not, then the message would be green if "joined" was present - to show a client joining - or red if "left" was present - to show a client disconnecting.